+++ module/data/actors/actor-character.mjs
@@ -52,6 +52,9 @@ export default class SWNCharacter extends SWNActorBase {
       quickSkill2: SWNShared.emptyString(), //deprecated
       quickSkill3: SWNShared.emptyString(), //deprecated
       extraHeader: SWNShared.emptyString(),
+      otherLabel: SWNShared.requiredString(game.i18n.localize("swnr.item.locationOther")),
+      extraLabel: SWNShared.requiredString(game.settings.get("swnr", "defaultExtraLabel")),
+      //extraLabel: SWNShared.requiredString(game.i18n.localize("swnr.item.locationExtra")),
       showResourceList: new fields.BooleanField({initial: false}),
       showCyberware: new fields.BooleanField({initial: true}),
       showPsychic: new fields.BooleanField({initial: true}),
@@ -310,6 +313,13 @@ export default class SWNCharacter extends SWNActorBase {
     
     // Calculate resource pools from Features/Foci/Edges
     this._calculateResourcePools();
+
+    this.locations = {
+      readied: game.i18n.localize("swnr.item.locationReadied"),
+      stowed: game.i18n.localize("swnr.item.locationStowed"),
+      other: this.tweak.otherLabel,
+      extra: this.tweak.extraLabel,
+    }
   }
 
   getRollData() {